will an int type display better?

+++ source/test-i2c/DeviceI2C.cpp
@@ -9,7 +9,7 @@ TEST_CASE(TestDeviceI2C) {
 
         device.write (0x43, 0x10);
         Pause::milli (10);
-        byte x = device.read (0x43);
+        int x = device.read (0x43);
         TEST_XY(x, 0x10);
         Pause::milli (10);
